Added tests to Market, back to BuyPhase

+++ EconomicCalculator/Storage/IMarket.cs
@@ -32,7 +32,7 @@ namespace EconomicCalculator.Storage
         /// <summary>
         /// The Territorial extent of the Market in acres. -1 means infinite.
         /// </summary>
-        int Territory { get; }
+        double Territory { get; }
 
         #endregion GeneralInfo
